feat: Implementation of Latitude/Longitude analysis
The major change was the implementation of the analysis for deciding whether keeping the variables pickup_latitude, pickup_longitude, dropoff_latitude and dropoff_longitude. Concluded that trip_duration is highly dependable on these variables, so they will be included on the model.

Also, considered the distance between pickup and dropoff for detecting
outliers. Distances over 125km were disconsidered.

Last but not least, reorganized the section, subsection, etc order